Working on Update Details pt1

+++ MediaBazaar/BusinessLogicLayer/EmployeeManager.cs
@@ -72,7 +72,7 @@ namespace MediaBazaar.Classes
             }
             employeeDAL.DeleteEmployee(employeeId);
             employees.Remove(employee);
-            _emergencyContactManager.RemoveEmergencyContactFromDB(employee.EmergencyContact.Id);
+            _emergencyContactManager.RemoveEmergencyContactFromDB((int)employee.EmergencyContact.Id);
         }
 
         public Employee GetEmployeeById(int employeeID)
@@ -80,7 +80,7 @@ namespace MediaBazaar.Classes
             return employees.FirstOrDefault(e => e.EmployeeID == employeeID);
         }
 
-        public List<Employee> GetAllRegularEmployees()
+		public List<Employee> GetAllRegularEmployees()
         {
             return employees.Where(e => !e.IsManager).ToList();
         }
